Cloud_server/backend/requirements.txt
Flask>=2.2
Werkzeug>=2.2
Flask-CORS==3.0.10
psycopg2-binary==2.9.3
flask-jwt-extended==4.7.1
bcrypt==4.1.3
flask-wtf==1.2.1
wtforms==3.2.1
flask-mail==0.10.0
geopy==2.4.1
pycountry==22.3.5
minio==7.1.16
overpy==0.6
requests>=2.31.0
pymongo==4.7.1
docker>=6.1.3